@@ -14,27 +14,34 @@
     "url": "git+https://github.com/awaludinar/unestorage"
   },
   "keywords": [
-    "NestJS"
+    "Unestorage",
+    "NestJS",
+    "unstorage"
   ],
   "author": "Awaludin AR <me@awaludin.dev>",
   "license": "MIT",
   "devDependencies": {
-    "@nestjs/common": "^11.1.2",
-    "@nestjs/core": "^11.1.2",
-    "@nestjs/platform-express": "^11.1.2",
-    "@nestjs/testing": "^11.1.2",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.36.0",
+    "@nestjs/common": "^11.1.6",
+    "@nestjs/core": "^11.1.6",
+    "@nestjs/platform-express": "^11.1.6",
+    "@nestjs/testing": "^11.1.6",
     "@trivago/prettier-plugin-sort-imports": "^5.2.2",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^22.15.23",
-    "eslint": "^9.27.0",
-    "jest": "^29.7.0",
-    "prettier": "^3.5.3",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.6.1",
+    "eslint": "^9.36.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
+    "jest": "^30.2.0",
+    "prettier": "^3.6.2",
     "rimraf": "^6.0.1",
-    "semantic-release": "^24.2.5",
-    "ts-jest": "^29.3.4",
-    "tsx": "^4.19.4",
-    "typescript-eslint": "^8.33.0",
-    "unstorage": "^1.16.0"
+    "semantic-release": "^24.2.9",
+    "ts-jest": "^29.4.4",
+    "tsx": "^4.20.6",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.45.0",
+    "unstorage": "^1.17.1"
   },
   "peerDependencies": {
     "@nestjs/common": "^10.0.0 || ^11.0.0",
